backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pdfplumber==0.10.3
spacy==3.7.2
scikit-learn==1.4.0
joblib==1.3.2
numpy==1.26.3
python-docx==1.1.0
python-dateutil==2.8.2
